Migrated ReflectionTools.
